requirements.txt
streamlit
pandas
requests
beautifulsoup4
yt-dlp
fastapi
uvicorn
httpx
altair
watchdog
python-dotenv